frontend merge

+++ server/src/main/java/com/codecool/DTO/location/LocationDTO.java
@@ -8,5 +8,5 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 
 //TODO use ContactInfoDTO, OpeningHoursDTO and EventDTO instead?
-public record LocationDTO(int id, String name, String address, String phone, String email, String description, @JsonIgnore UserDTO adminUser) {
+public record LocationDTO(long id, String name, String address, String phone, String email, String description, @JsonIgnore UserDTO adminUser) {
 }
